feat: renaming and hook
