+++ app/routes/index.tsx
@@ -3,8 +3,7 @@ import { Navigation } from "~/components/navigation";
 import { LogoBanner } from "~/components/logo-banner";
 import { i18n } from "~/i18n/i18n-utils";
 import { FeedbackDialog } from "~/components/feedback-dialog";
-import { useDialogStore } from "~/components/feedback-dialog/store/dialog";
-import { useEffect } from "react";
+import { useShowFeedbackDialogAfterCompletion } from "~/hooks/use-show-feedback-dialog-after-completion";
 
 export const meta = () => {
 	return [
@@ -18,19 +17,7 @@ export const meta = () => {
 };
 
 export default function Index() {
-	const { isOpen } = useDialogStore();
-
-	useEffect(() => {
-		const dialog = document.getElementById(
-			"feedback-dialog",
-		) as HTMLDialogElement;
-
-		if (!dialog || !isOpen) {
-			return;
-		}
-
-		dialog.showModal();
-	}, [isOpen]);
+	useShowFeedbackDialogAfterCompletion();
 
 	return (
 		<>
